provide efficient runtime for very large models
- openai API as a baseline
- pipeline callbacks
- docs
- document loader for RAG
- add complete()
- add streaming test
- implement REST API streaming

+++ omegaml/store/base.py
@@ -700,17 +700,17 @@ class OmegaStore(object):
             results.append((name, result))
         return dict(results) if report else len(results) > 0
 
-    def _drop(self, name, force=False, version=-1, **kwargs):
+    def _drop(self, name, force=False, version=-1, keep_data=False, **kwargs):
         meta = self.metadata(name, version=version)
         if meta is None and not force:
             raise DoesNotExist(name)
         collection = self.collection(name)
-        if collection:
+        if collection and not keep_data:
             self.mongodb.drop_collection(collection.name)
         if meta:
-            if meta.collection:
+            if meta.collection and not keep_data:
                 self.mongodb.drop_collection(meta.collection)
-            if meta and meta.gridfile is not None:
+            if meta and meta.gridfile is not None and not keep_data:
                 meta.gridfile.delete()
             self._drop_metadata(name)
             return True
